a HUGE cleanup for the garg code
