+++ ReturnOfTheAncients/Mono/Creatures/GargEssentials/GargantuanTentacleAttack.cs
@@ -1,19 +1,21 @@
-﻿using ECCLibrary;
-using System.Collections;
-using UnityEngine;
-
-namespace RotA.Mono.Creatures.GargEssentials
+﻿namespace RotA.Mono.Creatures.GargEssentials
 {
+    using ECCLibrary;
+    using System.Collections;
+    using UnityEngine;
+    
     public class GargantuanTentacleAttack : MeleeAttack
     {
-        private AudioSource attackSource;
-        private ECCAudio.AudioClipPool biteClipPool;
-        private GargantuanBehaviour behaviour;
-        private TrailManager leftTentacleTrail;
-        private TrailManager rightTentacleTrail;
+        GargantuanGrab grab;
+        AudioSource attackSource;
+        ECCAudio.AudioClipPool biteClipPool;
+        GargantuanBehaviour behaviour;
+        TrailManager leftTentacleTrail;
+        TrailManager rightTentacleTrail;
 
         void Start()
         {
+            grab = GetComponent<GargantuanGrab>();
             attackSource = gameObject.AddComponent<AudioSource>();
             attackSource.minDistance = 10f;
             attackSource.maxDistance = 40f;
@@ -32,7 +34,7 @@ namespace RotA.Mono.Creatures.GargEssentials
             {
                 return;
             }
-            if (liveMixin.IsAlive() && Time.time > behaviour.timeCanAttackAgain) //If it can attack, continue
+            if (liveMixin.IsAlive() && Time.time > grab.timeCanAttackAgain) //If it can attack, continue
             {
                 Creature thisCreature = gameObject.GetComponent<Creature>();
                 if (thisCreature.Aggression.Value >= 0.9f && thisCreature.Hunger.Value >= 0.6f) //This creature must be super angry to do this
@@ -42,7 +44,7 @@ namespace RotA.Mono.Creatures.GargEssentials
                     {
                         return;
                     }
-                    if (!behaviour.IsHoldingVehicle())
+                    if (!grab.IsHoldingVehicle())
                     {
                         Player player = target.GetComponent<Player>();
                         if (player != null)
@@ -52,19 +54,19 @@ namespace RotA.Mono.Creatures.GargEssentials
                                 return;
                             }
                         }
-                        else if (behaviour.GetCanGrabVehicle())
+                        else if (grab.GetCanGrabVehicle())
                         {
                             SeaMoth component4 = target.GetComponent<SeaMoth>();
                             if (component4 && !component4.docked)
                             {
-                                behaviour.GrabGenericSub(component4);
+                                grab.GrabGenericSub(component4);
                                 thisCreature.Aggression.Value -= 0.25f;
                                 return;
                             }
                             Exosuit component5 = target.GetComponent<Exosuit>();
                             if (component5 && !component5.docked)
                             {
-                                behaviour.GrabExosuit(component5);
+                                grab.GrabExosuit(component5);
                                 thisCreature.Aggression.Value -= 0.25f;
                                 return;
                             }
@@ -85,7 +87,7 @@ namespace RotA.Mono.Creatures.GargEssentials
                             if (liveMixin.health - num <= 0f) // make sure that the nodamage cheat is not on
                             {
                                 StartCoroutine(PerformBiteAttack(target));
-                                behaviour.timeCanAttackAgain = Time.time + 4f;
+                                grab.timeCanAttackAgain = Time.time + 4f;
                                 attackSource.clip = biteClipPool.GetRandomClip();
                                 attackSource.Play();
                             }
@@ -130,7 +132,7 @@ namespace RotA.Mono.Creatures.GargEssentials
         }
         public void OnVehicleReleased() //Called by gargantuan behavior. Gives a cooldown until the next bite.
         {
-            behaviour.timeCanAttackAgain = Time.time + 4f;
+            grab.timeCanAttackAgain = Time.time + 4f;
         }
         private IEnumerator PerformBiteAttack(GameObject target) //A delayed attack, to let him chomp down first.
         {
