+++ ReturnOfTheAncients/Mono/GargantuanEncounterPDA.cs
@@ -9,10 +9,10 @@ namespace RotA.Mono
     /// </summary>
     public class GargantuanEncounterPDA : MonoBehaviour
     {
-        private StoryGoal goal = new StoryGoal("GargantuanEncounter", Story.GoalType.Story, 0f);
+        StoryGoal goal = new StoryGoal("GargantuanEncounter", Story.GoalType.Story, 0f);
         public float maxDistance = 125f;
 
-        private void Start()
+        void Start()
         {
             if (StoryGoalManager.main.IsGoalComplete(goal.key))
             {
@@ -28,7 +28,7 @@ namespace RotA.Mono
         {
             if (Vector3.Distance(transform.position, Player.main.transform.position) < maxDistance)
             {
-                if (!GargantuanBehaviour.PlayerInPrecursorBase())
+                if (!GragantuanConditions.PlayerInPrecursorBase())
                 {
                     if (StoryGoalManager.main.OnGoalComplete(goal.key))
                     {
